Everything except chat is a popup window now
