+++ src/ui/popups/mod.rs
@@ -1,0 +1,63 @@
+pub mod help;
+
+use tui::layout::{Constraint, Direction, Layout, Rect};
+
+pub enum Centering {
+    Input(u16, u16),
+    Box(u16, u16),
+}
+
+pub fn centered_popup(c: Centering, size: Rect) -> Rect {
+    match c {
+        Centering::Box(percent_x, percent_y) => {
+            let popup_layout = Layout::default()
+                .direction(Direction::Vertical)
+                .constraints(
+                    [
+                        Constraint::Percentage((100 - percent_y) / 2),
+                        Constraint::Percentage(percent_y),
+                        Constraint::Percentage((100 - percent_y) / 2),
+                    ]
+                    .as_ref(),
+                )
+                .split(size);
+
+            Layout::default()
+                .direction(Direction::Horizontal)
+                .constraints(
+                    [
+                        Constraint::Percentage((100 - percent_x) / 2),
+                        Constraint::Percentage(percent_x),
+                        Constraint::Percentage((100 - percent_x) / 2),
+                    ]
+                    .as_ref(),
+                )
+                .split(popup_layout[1])[1]
+        }
+        Centering::Input(percent_x, percent_y) => {
+            let popup_layout = Layout::default()
+                .direction(Direction::Vertical)
+                .constraints(
+                    [
+                        Constraint::Percentage((100 - percent_y) / 2),
+                        Constraint::Length(3),
+                        Constraint::Percentage((100 - percent_y) / 2),
+                    ]
+                    .as_ref(),
+                )
+                .split(size);
+
+            Layout::default()
+                .direction(Direction::Horizontal)
+                .constraints(
+                    [
+                        Constraint::Percentage((100 - percent_x) / 2),
+                        Constraint::Percentage(percent_x),
+                        Constraint::Percentage((100 - percent_x) / 2),
+                    ]
+                    .as_ref(),
+                )
+                .split(popup_layout[1])[1]
+        }
+    }
+}
